[MS] Apply filters to invoices page
